Added CSS and fiddled with the file layout
I changed the file layout and added styling to the index.

Added a cohesive colour-scheme and font family.
The CSS is very rough and does not adjust well to changes of window size. I will be fixing this in the future.
I will also plan to clean up the code to be more readable and dry.