Polishing up the CSS and HTML
I cleaned up my code a bit, though it can always be better. I will be cleaning it up further later.

I changed the shape of the search bar and search button and added a background colour transition for when it is clicked.

I have run into an issue where I am unable to change the colour of the text in the search bar. I am able to manipulate its weight or font size and other font related variables, but cannot manipulate the colour. I have inspected the code on the webpage. The inspect function displays the text as being the colour I chose but in reality the text remains a grey colour.

ah well. I shall figure it out when I next have time. When that time comes I will also be fixing the netlify link. Not now.