Updated the searchbar CSS
I updated the colour of the text in the search bar when text is entered

I'm still not sure how to change the placeholder text colour, but that can be changed a bit later. For now, I want to go ahead and finish this project after a hiatus from extra hours at work.